Remove hex27 and reorganize mesh reader.
