+++ tests/unit-tests/serial.cmake
@@ -232,7 +232,6 @@ specfem_add_test(mesh_dim3_tests
             mesh/dim3/boundaries.cpp
             mesh/dim3/adjacency_graph.cpp
             mesh/dim3/tags.cpp
-            mesh/dim3/globe_hex27.cpp
             mesh/dim3/globe_reader.cpp
             mesh/dim3/test.cpp
   LIBRARIES gtest_main
